requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
dvc==3.30.1
mlflow==2.15.1
pytest==7.4.0
pyyaml==6.0
evidently==0.4.3